put UUID functions in library
